Cache recursive interfaces
