+++ src/tools/illink/src/linker/Linker.Steps/MarkStep.cs
@@ -35,6 +35,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using System.Reflection.Metadata.Ecma335;
 using System.Reflection.Runtime.TypeParsing;
 using System.Text.RegularExpressions;
 using ILLink.Shared;
@@ -2449,10 +2450,9 @@ namespace Mono.Linker.Steps
 
 		void MarkInterfaceImplementations (TypeDefinition type)
 		{
-			if (!type.HasInterfaces)
+			var ifaces = Annotations.GetRecursiveInterfaces (type);
+			if (ifaces is null)
 				return;
-
-			var ifaces = GetRecursiveInterfaceImplementations (type);
 			foreach (var (ifaceType, impls) in ifaces) {
 				// Only mark interface implementations of interface types that have been marked.
 				// This enables stripping of interfaces that are never used
@@ -2461,89 +2461,6 @@ namespace Mono.Linker.Steps
 			}
 		}
 
-		List<(TypeReference, List<InterfaceImplementation>)> GetRecursiveInterfaceImplementations (TypeDefinition type)
-		{
-			List<(TypeReference, List<InterfaceImplementation>)> firstImplementationChain = new ();
-
-			AddRecursiveInterfaces (type, [], firstImplementationChain, Context);
-			Debug.Assert (firstImplementationChain.All (kvp => Context.Resolve (kvp.Item1) == Context.Resolve (kvp.Item2.Last ().InterfaceType)));
-
-			return firstImplementationChain;
-
-			static void AddRecursiveInterfaces (TypeReference typeRef, IEnumerable<InterfaceImplementation> pathToType, List<(TypeReference, List<InterfaceImplementation>)> firstImplementationChain, LinkContext Context)
-			{
-				var type = Context.TryResolve (typeRef);
-				if (type is null)
-					return;
-				// GeVt all explicit interfaces of this type
-				foreach (var directIface in type.Interfaces) {
-					//var directlyImplementedType = Context.Resolve (directIface.InterfaceType);
-					var directlyImplementedType = directIface.InterfaceType.TryInflateFrom (typeRef, Context);
-					if (directlyImplementedType is null) {
-						continue;
-					}
-					if (!firstImplementationChain.Any (i => InterfaceTypeEquals(i.Item1, directlyImplementedType, Context))) {
-						firstImplementationChain.Add ((directlyImplementedType, pathToType.Append (directIface).ToList ()));
-					}
-				}
-
-				// Add interfaces on base type
-				if (type.BaseType is { } baseType && Context.Resolve (baseType) is { } baseDef) {
-					AddRecursiveInterfaces (baseDef, pathToType, firstImplementationChain, Context);
-				}
-
-				// Recursive interfaces next to preserve Inherit/Implement tree order
-				foreach (var directIface in type.Interfaces) {
-					// If we can't resolve the interface type we can't find recursive interfaces
-					var ifaceDirectlyOnType = directIface.InterfaceType.TryInflateFrom(typeRef, Context);
-					if (ifaceDirectlyOnType is null) {
-						continue;
-					}
-					AddRecursiveInterfaces (ifaceDirectlyOnType, pathToType.Append (directIface), firstImplementationChain, Context);
-				}
-			}
-			/// <summary>
-			/// Compares two TypeReferences to interface types and determines if they are equivalent references, taking into account generic arguments and element types.
-			/// </summary>
-			static bool InterfaceTypeEquals (TypeReference? type, TypeReference? other, ITryResolveMetadata resolver)
-			{
-				Debug.Assert (type is not null && other is not null);
-				Debug.Assert (resolver.TryResolve (type)?.IsInterface is null or true);
-				Debug.Assert (resolver.TryResolve (other)?.IsInterface is null or true);
-				return TypeEquals (type, other);
-
-				bool TypeEquals (TypeReference type1, TypeReference type2)
-				{
-					if (type1 == type2)
-						return true;
-
-					if (resolver.TryResolve (type1) != resolver.TryResolve (type2))
-						return false;
-
-					if (type1 is GenericInstanceType genericInstance1) {
-						if (type2 is not GenericInstanceType genericInstance2)
-							return false;
-						if (genericInstance1.HasGenericParameters != genericInstance2.HasGenericParameters)
-							return false;
-						if (genericInstance1.GenericParameters.Count != genericInstance2.GenericParameters.Count
-							|| genericInstance2.GenericArguments.Count != genericInstance2.GenericArguments.Count)
-							return false;
-						for (var i = 0; i < genericInstance1.GenericArguments.Count; ++i) {
-							if (!TypeEquals (genericInstance1.GenericArguments[i], genericInstance2.GenericArguments[i]))
-								return false;
-						}
-						return true;
-					}
-
-					if (type1 is TypeSpecification typeSpec1) {
-						if (type2 is not TypeSpecification typeSpec2)
-							return false;
-						return TypeEquals (typeSpec1.ElementType, typeSpec2.ElementType);
-					}
-					return type1.FullName == type2.FullName;
-				}
-			}
-		}
 
 		protected virtual bool ShouldMarkInterfaceImplementationList (TypeDefinition type, List<InterfaceImplementation> ifaces, TypeReference ifaceType)
 		{
